regular

+++ src/main/java/com/example/orbix_web/controllers/RtvDetailServiceController.java
@@ -18,6 +18,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.example.orbix_web.exceptions.InvalidOperationException;
 import com.example.orbix_web.exceptions.NotFoundException;
 import com.example.orbix_web.exceptions.OperationFailedException;
 import com.example.orbix_web.models.Rtv;
@@ -39,6 +40,18 @@ public class RtvDetailServiceController {
 	@Autowired
 	RtvDetailRepository rtvDetailRepository;
 	
+	// get rtv detail
+	@Transactional
+	@RequestMapping(method = RequestMethod.GET, value = "/rtv_details/{id}")
+	public RtvDetail getRtvDetail(@PathVariable(value = "id") Long id) {
+		RtvDetail rtvDetail = null;
+		try {
+			rtvDetail = rtvDetailRepository.findById(id).get();
+		}catch(Exception e) {
+			throw new NotFoundException("RTV detail not found");
+		}
+		return rtvDetail;
+	}
 	// Create Rtv Detail
 	@Transactional
 	@RequestMapping(method = RequestMethod.GET, value = "/rtv_details/rtv_id= {rtv_id}")
@@ -57,7 +70,11 @@ public class RtvDetailServiceController {
     @Transactional
 	public boolean createRtvDetail(@Valid @RequestBody RtvDetail rtvDetail) {
 		boolean created = false;
-		Rtv rtv;
+		Rtv rtv = rtvRepository.findById(rtvDetail.getRtv().getId())
+                .orElseThrow(() -> new NotFoundException("RTV not found"));
+		if(rtv.getStatus().equals("COMPLETED")) {
+			throw new InvalidOperationException("Could not modify, RTV already completed");
+		}
     	try {
     		Long rtvId = (rtvDetail.getRtv().getId());
     		rtv = rtvRepository.findById(rtvId).get();
@@ -76,6 +93,11 @@ public class RtvDetailServiceController {
     @Transactional
 	public boolean updateRtvDetail(@PathVariable(value = "id") Long rtvDetailId, @Valid @RequestBody RtvDetail rtvDetail) {
 		boolean updated = false;
+		Rtv rtv = rtvRepository.findById(rtvDetail.getRtv().getId())
+                .orElseThrow(() -> new NotFoundException("RTV not found"));
+		if(rtv.getStatus().equals("COMPLETED")) {
+			throw new InvalidOperationException("Could not update, RTV already completed");
+		}
 		try {
 			rtvDetailRepository.saveAndFlush(rtvDetail);
 			updated = true;
@@ -89,7 +111,12 @@ public class RtvDetailServiceController {
 	public boolean deleteRtvDetail(@PathVariable(value = "id") Long rtvDetailId) {
 		boolean deleted = false;
 		RtvDetail rtvDetail = rtvDetailRepository.findById(rtvDetailId)
+                .orElseThrow(() -> new NotFoundException("RTV detail not found"));
+		Rtv rtv = rtvRepository.findById(rtvDetail.getRtv().getId())
                 .orElseThrow(() -> new NotFoundException("RTV not found"));
+		if(rtv.getStatus().equals("COMPLETED")) {
+			throw new InvalidOperationException("Could not delete, RTV already completed");
+		}
 		try {
 			rtvDetailRepository.delete(rtvDetail);
 			deleted = true;
